feat: remove markdown-it
