+++ example/content/lorem-ipsum-5.md
@@ -1,5 +1,5 @@
 ---
-title: Lorem Ipsum 5
+title: Our Supported Markdown
 date: 2024-04-20
 author: 'Humming Admin'
 categories: ["test"]
@@ -7,55 +7,40 @@ slug: lorem-ipsum-5
 image: '/img/bun-wikipedia.jpeg'
 draft: false
 ---
-Lorem 5 ipsum dolor sit amet, consectetur adipiscing elit. Curabitur ac auctor lectus, faucibus accumsan nisi. Curabitur sed porttitor augue, id accumsan risus. Ut commodo felis odio, sed cursus dolor ornare in. In quis tellus nulla. Quisque et enim leo. Donec ut ante et felis vehicula vestibulum aliquam eget nulla. In mollis laoreet felis, sit amet aliquam ipsum fermentum tincidunt. 
 
-Nam ex nulla, interdum ac orci ut, pulvinar semper enim. Praesent vulputate nisl id nisl hendrerit, sit amet vehicula nibh fermentum. Fusce vulputate enim quis ligula ultrices interdum. In quis sem est. Quisque consectetur fringilla nisi, ut luctus mi ultrices at.
+Here's our supported Markdown
 
-## Typographic replacements
-
-Enable typographer option to see result.
-
-(c) (C) (r) (R) (tm) (TM) (p) (P) +-
-
-test.. test... test..... test?..... test!....
-
-!!!!!! ???? ,,  -- ---
-
-"Smartypants, double quotes" and 'single quotes'
-
+# heading 1
+## heading 2
+### heading 3
+#### heading 4
+##### heading 5
+###### heading 6
 
 ## Emphasis
 
-**This is bold text**
+bold :
+- **This is bold text**
 
-__This is bold text__
+italic:
+- *This is italic text*
 
-*This is italic text*
-
-_This is italic text_
-
-~~Strikethrough~~
+strikethrough:
+- ~~Strikethrough~~
 
 
 ## Blockquotes
 
 
 > Blockquotes can also be nested...
->> ...by using additional greater-than signs right next to each other...
-> > > ...or with spaces between arrows.
-
 
 ## Lists
 
 Unordered
 
-+ Create a list by starting a line with `+`, `-`, or `*`
-+ Sub-lists are made by indenting 2 spaces:
-  - Marker character change forces new list start:
-    * Ac tristique libero volutpat at
-    + Facilisis in pretium nisl aliquet
-    - Nulla volutpat aliquam velit
-+ Very easy!
+- Create a list by starting a line with `-`
+- Sub-lists are made by indenting 2 spaces:
+- Very easy!
 
 Ordered
 
@@ -63,35 +48,17 @@ Ordered
 2. Consectetur adipiscing elit
 3. Integer molestie lorem at massa
 
-
-1. You can use sequential numbers...
-1. ...or keep all the numbers as `1.`
-
-Start numbering with offset:
-
-57. foo
-1. bar
-
-
 ## Code
 
-Inline `code`
+Inline `code` 
 
-Indented code
-
-    // Some comments
-    line 1 of code
-    line 2 of code
-    line 3 of code
-
-
-Block code "fences"
+## block code
 
 ```
-Sample text here...
+Block code
 ```
 
-Syntax highlighting
+## Syntax highlighting
 
 ``` js
 var foo = function (bar) {
@@ -101,129 +68,10 @@ var foo = function (bar) {
 console.log(foo(5));
 ```
 
-## Tables
-
-| Option | Description |
-| ------ | ----------- |
-| data   | path to data files to supply the data that will be passed into templates. |
-| engine | engine to be used for processing templates. Handlebars is the default. |
-| ext    | extension to be used for dest files. |
-
-Right aligned columns
-
-| Option | Description |
-| ------:| -----------:|
-| data   | path to data files to supply the data that will be passed into templates. |
-| engine | engine to be used for processing templates. Handlebars is the default. |
-| ext    | extension to be used for dest files. |
-
-
 ## Links
 
 [link text](http://dev.nodeca.com)
 
-[link with title](http://nodeca.github.io/pica/demo/ "title text!")
-
-Autoconverted link https://github.com/nodeca/pica (enable linkify to see)
-
-
 ## Images
 
 ![Minion](https://octodex.github.com/images/minion.png)
-![Stormtroopocat](https://octodex.github.com/images/stormtroopocat.jpg "The Stormtroopocat")
-
-Like links, Images also have a footnote style syntax
-
-![Alt text][id]
-
-With a reference later in the document defining the URL location:
-
-[id]: https://octodex.github.com/images/dojocat.jpg  "The Dojocat"
-
-
-## Plugins
-
-The killer feature of `markdown-it` is very effective support of
-[syntax plugins](https://www.npmjs.org/browse/keyword/markdown-it-plugin).
-
-
-### [Emojies](https://github.com/markdown-it/markdown-it-emoji)
-
-> Classic markup: :wink: :cry: :laughing: :yum:
->
-> Shortcuts (emoticons): :-) :-( 8-) ;)
-
-see [how to change output](https://github.com/markdown-it/markdown-it-emoji#change-output) with twemoji.
-
-
-### [Subscript](https://github.com/markdown-it/markdown-it-sub) / [Superscript](https://github.com/markdown-it/markdown-it-sup)
-
-- 19^th^
-- H~2~O
-
-
-### [\<ins>](https://github.com/markdown-it/markdown-it-ins)
-
-++Inserted text++
-
-
-### [\<mark>](https://github.com/markdown-it/markdown-it-mark)
-
-==Marked text==
-
-
-### [Footnotes](https://github.com/markdown-it/markdown-it-footnote)
-
-Footnote 1 link[^first].
-
-Footnote 2 link[^second].
-
-Inline footnote^[Text of inline footnote] definition.
-
-Duplicated footnote reference[^second].
-
-[^first]: Footnote **can have markup**
-
-    and multiple paragraphs.
-
-[^second]: Footnote text.
-
-
-### [Definition lists](https://github.com/markdown-it/markdown-it-deflist)
-
-Term 1
-
-:   Definition 1
-with lazy continuation.
-
-Term 2 with *inline markup*
-
-:   Definition 2
-
-        { some code, part of Definition 2 }
-
-    Third paragraph of definition 2.
-
-_Compact style:_
-
-Term 1
-  ~ Definition 1
-
-Term 2
-  ~ Definition 2a
-  ~ Definition 2b
-
-
-### [Abbreviations](https://github.com/markdown-it/markdown-it-abbr)
-
-This is HTML abbreviation example.
-
-It converts "HTML", but keep intact partial entries like "xxxHTMLyyy" and so on.
-
-*[HTML]: Hyper Text Markup Language
-
-### [Custom containers](https://github.com/markdown-it/markdown-it-container)
-
-::: warning
-*here be dragons*
-:::
